feat: enable downloads for markdown and docx
