+++ guillermo-lam-cv/src/App.tsx
@@ -3,6 +3,7 @@ import { createSignal } from "solid-js";
 import MarkdownIt from "markdown-it";
 import LangSwitcher from "./components/LangSwitcher";
 import DownloadButtons from "./components/DownloadButtons";
+import { getLocaleParts } from "./markdown";
 
 // --- Types and helpers ---
 type Locale = "en" | "es" | "fr" | "de";
@@ -16,40 +17,6 @@ function detectLocale(): Locale {
   return LOCALES.includes(nav) ? nav : FALLBACK;
 }
 
-// --- Eager import of ALL markdown files in all locales & sections ---
-const mdModules = import.meta.glob("./parts/*/*/*.md", {
-  eager: true,
-  query: "?raw",
-  import: "default",
-});
-const headerModules = import.meta.glob("./parts/*/header_summary.md", {
-  eager: true,
-  query: "?raw",
-  import: "default",
-});
-
-// --- Compose Markdown in desired order ---
-function getLocaleParts(locale: Locale) {
-  // 1. Header
-  const header = headerModules[`./parts/${locale}/header_summary.md`] || "";
-  // 2. Education
-  const education = [
-    mdModules[`./parts/${locale}/education/education.md`] || "",
-    mdModules[`./parts/${locale}/education/certifications.md`] || "",
-  ].join("\n\n");
-  // 3. Volunteering
-  const volunteering =
-    mdModules[`./parts/${locale}/volunteering/volunteering.md`] || "";
-  // 4. Work: all work/experience_*.md files, sorted for stability
-  const workKeys = Object.keys(mdModules)
-    .filter((k) => k.startsWith(`./parts/${locale}/work/`) && k.endsWith(".md"))
-    .sort();
-  const work = workKeys.map((k) => mdModules[k]).join("\n\n");
-  // --- Stitch together ---
-  return [header, education, volunteering, work]
-    .filter(Boolean)
-    .join("\n\n---\n\n");
-}
 
 export default function App() {
   const [locale, setLocale] = createSignal<Locale>(detectLocale());
